Reconcile StartingSyncPivotUpdaterTests with the master refactor that moved the
updater from the ISyncModeSelector.Changed event model to the method-based
EnsureSyncPivot(CancellationToken) model: new constructor (drops
ISyncModeSelector, adds ISyncProgressResolver), RlpReader/DecodeULong, and ulong
pivot numbers. Kept this PR's regression tests for the block-tree finalized-hash
fallback, adapted to the new async API.

+++ src/Nethermind/Nethermind.Benchmark/State/ReadOnlySnapshotBundleBenchmark.cs
@@ -66,7 +66,7 @@ public class ReadOnlySnapshotBundleBenchmark
         // Track storage account ranges per snapshot for hit distribution
         List<(int AddressStart, int StorageCount, int SlotsPerAccount)> storageRanges = [];
 
-        for (int block = 0; block < SnapshotCount; block++)
+        for (ulong block = 0; block < SnapshotCount; block++)
         {
             int multiplier = block < 6 ? 16 : 1;
             int accountCount = 1000 * multiplier;
